Log from install.cmd with date and time

+++ dotnet/content/install.ps1
@@ -2,10 +2,13 @@
 # https://github.com/projectkudu/kudu/blob/13824205c60a4bdb53896b9553ef1f370a93b911/Kudu.Services/SiteExtensions/SiteExtensionController.cs#L240
 
 # Set the unique pipe names in the applicationHost.xdt file for traces and metrics
-$tracePipeId=([guid]::NewGuid().ToString().ToUpper())
-$statsPipeId=([guid]::NewGuid().ToString().ToUpper())
-((Get-Content -path .\applicationHost.xdt.dd -Raw) -replace "uniqueTracePipeId", "${tracePipeId}") | Set-Content -Path .\applicationHost.xdt.dd
-((Get-Content -path .\applicationHost.xdt.dd -Raw) -replace "uniqueStatsPipeId", "${statsPipeId}") | Set-Content -Path .\applicationHost.xdt.dd
+if ([System.IO.File]::Exists('.\applicationHost.xdt.dd')) 
+{
+  $tracePipeId=([guid]::NewGuid().ToString().ToUpper())
+  $statsPipeId=([guid]::NewGuid().ToString().ToUpper())
+  ((Get-Content -path .\applicationHost.xdt.dd -Raw) -replace "uniqueTracePipeId", "${tracePipeId}") | Set-Content -Path .\applicationHost.xdt.dd
+  ((Get-Content -path .\applicationHost.xdt.dd -Raw) -replace "uniqueStatsPipeId", "${statsPipeId}") | Set-Content -Path .\applicationHost.xdt.dd
+}
 
 # If a web application is running, do not create the applicationHost.xdt file
 # This is used as an indication of install failure
